requirements.txt
beautifulsoup4==4.9.3
mongoengine==0.23.0
python-dotenv==0.17.1
requests==2.25.1
Shapely==1.7.1
twilio==6.57.0
